📝 fix(../src/pages): rename page to pages folder
